requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-tenants==3.5.0
django-cors-headers==4.3.1
django-environ==0.12.0
drf-yasg==1.21.7
channels==4.0.0
daphne==4.0.0
celery==5.3.4
redis==5.0.1
channels-redis==4.1.0
psycopg2-binary==2.9.10
Pillow==10.2.0
openpyxl==3.1.2
reportlab==4.0.7
requests==2.31.0
djangorestframework-simplejwt==5.3.0 